requirements.txt
flask
python-dotenv
langchain
langchain-openai
langchain-community
faiss-cpu
huggingface-hub
docx2txt
unstructured
python-docx
